change argument types of wfd_sink_start_session()
To workaround valac issue

+++ src/ctl/ctl.h
@@ -174,10 +174,10 @@ int wfd_out_session_new(struct wfd_session **out,
 				struct wfd_sink *sink,
 				const char *authority,
 				const char *display,
-				uint16_t x,
-				uint16_t y,
-				uint16_t width,
-				uint16_t height,
+				uint32_t x,
+				uint32_t y,
+				uint32_t width,
+				uint32_t height,
 				const char *audio_dev);
 int wfd_session_start(struct wfd_session *s, uint64_t id);
 enum wfd_session_dir wfd_session_get_dir(struct wfd_session *s);
@@ -222,10 +222,10 @@ int wfd_sink_start_session(struct wfd_sink *sink,
 				struct wfd_session **out,
 				const char *authority,
 				const char *display,
-				uint16_t x,
-				uint16_t y,
-				uint16_t width,
-				uint16_t height,
+				uint32_t x,
+				uint32_t y,
+				uint32_t width,
+				uint32_t height,
 				const char *audio_dev);
 void wfd_sink_handle_session_ended(struct wfd_sink *sink);
 bool wfd_sink_is_session_started(struct wfd_sink *sink);
