+++ src/disp/wfd-out-session.c
@@ -69,10 +69,10 @@ int wfd_out_session_new(struct wfd_session **out,
 				struct wfd_sink *sink,
 				const char *authority,
 				const char *display,
-				uint16_t x,
-				uint16_t y,
-				uint16_t width,
-				uint16_t height,
+				uint32_t x,
+				uint32_t y,
+				uint32_t width,
+				uint32_t height,
 				const char *audio_dev)
 {
 	_shl_free_ char *display_schema = NULL;
