+++ src/disp/wfd-sink.c
@@ -121,10 +121,10 @@ int wfd_sink_start_session(struct wfd_sink *sink,
 				struct wfd_session **out,
 				const char *authority,
 				const char *display,
-				uint16_t x,
-				uint16_t y,
-				uint16_t width,
-				uint16_t height,
+				uint32_t x,
+				uint32_t y,
+				uint32_t width,
+				uint32_t height,
 				const char *audio_dev)
 {
 	int r;
